Upgrade int32 to int64

+++ volpe/builder.py
@@ -6,7 +6,7 @@ from llvmlite import ir
 from builder_utils import write_environment, free_environment, options, \
     read_environment, tuple_assign, copy, copy_environment, build_closure, free
 from tree import TypeTree
-from volpe_types import int1, flt64, VolpeClosure, int32, target_data
+from volpe_types import int1, int32, int64, flt64, VolpeClosure, target_data
 
 
 class LLVMScope(Interpreter):
@@ -132,7 +132,7 @@ class LLVMScope(Interpreter):
         length = self.builder.extract_value(list_value, 1)
 
         before_end = self.builder.icmp_signed("<", i, length)
-        more_than_0 = self.builder.icmp_signed(">=", i, int32(0))
+        more_than_0 = self.builder.icmp_signed(">=", i, int64(0))
         in_range = self.builder.and_(before_end, more_than_0)
         with self.builder.if_then(self.builder.not_(in_range)):
             self.builder.unreachable()
@@ -161,7 +161,7 @@ class LLVMScope(Interpreter):
         with self.builder.if_then(self.builder.icmp_signed("<", phi, length)):
             # TODO fix lists
             # self.builder.store(closure_call(self.builder, closure, [phi]), self.builder.gep(pointer, [phi]))
-            ret(self.builder.add(phi, int32(1)))
+            ret(self.builder.add(phi, int64(1)))
 
         free(self.builder, closure)
 
@@ -291,7 +291,7 @@ class LLVMScope(Interpreter):
 
     def convert_flt(self, tree: TypeTree):
         value = self.visit_children(tree)[0]
-        return self.builder.fptosi(value, int32)
+        return self.builder.fptosi(value, int64)
 
     def equals_flt(self, tree: TypeTree):
         values = self.visit_children(tree)
